requirements.txt
amqp==2.1.4
appdirs==1.4.3
billiard==3.5.0.2
celery==4.0.2
cmsplugin-filer==1.1.3
dj-database-url==0.4.2
Django==1.10.6
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cms==3.4.4
django-ckeditor==5.3.0
django-filer==1.2.7
django-formtools==2.0
django-mptt==0.8.7
django-polymorphic==1.0.2
django-rosetta==0.7.13
django-sekizai==0.10.0
django-treebeard==4.1.0
djangocms-admin-style==1.2.6.2
djangocms-attributes-field==0.2.0
djangocms-column==1.7.0
djangocms-googlemap==1.0.2
djangocms-installer==0.9.6
djangocms-link==2.1.1
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.4.0
djangocms-video==2.0.3
django-hvad==1.8.0
django-debug-toolbar==1.8
django-pylibmc==0.6.1
easy-thumbnails==2.3
html5lib==0.9999999
kombu==4.0.2
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
tzlocal==1.3
Unidecode==0.4.20
vine==1.1.3
whitenoise==3.3.0
dnspython==1.15.0
pylibmc==1.5.2
sqlparse==0.2.3

git+https://github.com/benoitc/gunicorn.git#egg=gunicorn
